requirements.txt
anyio==3.6.1
bidict==0.22.0
click==8.1.3
dnspython==1.16.0
eventlet==0.30.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
python-dotenv==0.20.0
python-engineio==4.3.3
python-socketio==5.7.1
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
